fix: changed social banners width from "dv-main"
all the "a" tags width were set to 100% which meant that when hovering over the div the user could select the link (which is not supposed to be the case). Decreased the width of the "a" tags and set the alignment of the "sc-social-banners" to the end.# Please enter the commit message for your changes. Lines starng